+++ src/main/resources/templates/static.html
