+++ src/test/kotlin/ecommerce/api/ProductControllerTest.kt
@@ -1,5 +1,8 @@
 package ecommerce.api
 
+import ecommerce.controller.api.ProductController
+import ecommerce.model.Product
+import ecommerce.service.ProductService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
