+++ src/test/kotlin/ecommerce/api/ProductServiceTest.kt
@@ -1,5 +1,7 @@
 package ecommerce.api
 
+import ecommerce.model.Product
+import ecommerce.service.ProductService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
